embeddings_service/pyproject.toml
[tool.poetry]
name = "embeddings-service"
version = "0.1.0"
description = ""
authors = ["Henry Merrilees <hmerrilees@hmc.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.0"
uvicorn = "^0.23.2"
sentence-transformers = "^2.2.2"
grpcio = "^1.57.0"
grpcio-tools = "^1.57.0"
protobuf = "^4.24.0"
typer = "^0.9.0"
rich = "^13.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
